refact: split and separate action buttons
