+++ app/dashboard/src/pages/users/table/index.tsx
@@ -22,11 +22,8 @@ import {
 import classNames from 'classnames';
 
 import { statusColors } from 'constants/Settings';
-import { t } from 'i18next';
 import { FC, Fragment, useEffect, useState } from 'react';
-import CopyToClipboard from 'react-copy-to-clipboard';
 import { useTranslation } from 'react-i18next';
-import { User } from 'types';
 import { OnlineBadge } from 'components/online-badge';
 import { OnlineStatus } from 'components/online-status';
 import { StatusBadge } from 'components/status-badge';
@@ -35,15 +32,12 @@ import {
   EmptySection,
   AccordionArrowIcon,
   EditIcon,
-  CopiedIcon,
-  CopyIcon,
   StatusSortSelect,
   Sort,
-  QRIcon,
-  SubscriptionLinkIcon
 } from 'components/table';
 import { UsageSlider, UsageSliderCompact } from './usage-slider';
 import { useUsers } from 'stores';
+import { ActionButtons } from './action-buttons';
 
 type ExpandedIndex = number | number[];
 
@@ -446,132 +440,5 @@ export const UsersTable: FC<UsersTableProps> = (props) => {
       </Table>
       <Pagination filters={filters} total={total} onFilterChange={onFilterChange} />
     </Box>
-  );
-};
-
-type ActionButtonsProps = {
-  user: User;
-};
-
-const ActionButtons: FC<ActionButtonsProps> = ({ user }) => {
-  const { setQRCode, setSubLink } = useUsers();
-
-  const proxyLinks = user.links.join('\r\n');
-
-  const [copied, setCopied] = useState([-1, false]);
-  useEffect(() => {
-    if (copied[1]) {
-      setTimeout(() => {
-        setCopied([-1, false]);
-      }, 1000);
-    }
-  }, [copied]);
-  return (
-    <HStack
-      justifyContent="flex-end"
-      onClick={(e) => {
-        e.preventDefault();
-        e.stopPropagation();
-      }}
-    >
-      <CopyToClipboard
-        text={
-          user.subscription_url.startsWith('/')
-            ? window.location.origin + user.subscription_url
-            : user.subscription_url
-        }
-        onCopy={() => {
-          setCopied([0, true]);
-        }}
-      >
-        <div>
-          <Tooltip
-            label={
-              copied[0] == 0 && copied[1]
-                ? t('usersTable.copied')
-                : t('usersTable.copyLink')
-            }
-            placement="top"
-          >
-            <IconButton
-              p="0 !important"
-              aria-label="copy subscription link"
-              bg="transparent"
-              _dark={{
-                _hover: {
-                  bg: 'gray.700',
-                },
-              }}
-              size={{
-                base: 'sm',
-                md: 'md',
-              }}
-            >
-              {copied[0] == 0 && copied[1] ? (
-                <CopiedIcon />
-              ) : (
-                <SubscriptionLinkIcon />
-              )}
-            </IconButton>
-          </Tooltip>
-        </div>
-      </CopyToClipboard>
-      <CopyToClipboard
-        text={proxyLinks}
-        onCopy={() => {
-          setCopied([1, true]);
-        }}
-      >
-        <div>
-          <Tooltip
-            label={
-              copied[0] == 1 && copied[1]
-                ? t('usersTable.copied')
-                : t('usersTable.copyConfigs')
-            }
-            placement="top"
-          >
-            <IconButton
-              p="0 !important"
-              aria-label="copy configs"
-              bg="transparent"
-              _dark={{
-                _hover: {
-                  bg: 'gray.700',
-                },
-              }}
-              size={{
-                base: 'sm',
-                md: 'md',
-              }}
-            >
-              {copied[0] == 1 && copied[1] ? <CopiedIcon /> : <CopyIcon />}
-            </IconButton>
-          </Tooltip>
-        </div>
-      </CopyToClipboard>
-      <Tooltip label="QR Code" placement="top">
-        <IconButton
-          p="0 !important"
-          aria-label="qr code"
-          bg="transparent"
-          _dark={{
-            _hover: {
-              bg: 'gray.700',
-            },
-          }}
-          size={{
-            base: 'sm',
-            md: 'md',
-          }}
-          onClick={() => {
-            setQRCode(user.links);
-            setSubLink(user.subscription_url);
-          }}
-        >
-          <QRIcon />
-        </IconButton>
-      </Tooltip>
-    </HStack >
   );
 };
